vo_gpu: hwdec_cuda: Implement interop for placebo
This change updates the vulkan interop code to work with the
libplacebo based ra_vk, but also introduces the semaphore based
synchronisation I've already been working on to replace the
buffer pool.

Synchronisation means we can safely reuse the same buffer for each
mapped frame, by ensuring the buffer is copied to the texture before
mapping the next frame.