WIP: vo_gpu: hwdec_vaegl: vulkan interop
This change is a first pass at vulkan interop for vaapi. I took the
existing hwdec_vaegl and replaced all the EGL logic with equivalent
Vulkan logic. Whether we end up with a separate file or separate
logic in one file (like with hwdec_cuda), I will need to adjust it
relative to what is in this change. So it's very much WIP.

The basic logic is very close to what is done for EGL except that
I've wrapped most of the fiddly bits in libplacebo so there is less
to see here. I also threw out the support for the older libva to
make things easier to read but it's still possible to support the
old API if desired.

The only really big item of note is that this is still using split
planes rather than attempting to use a multiplane surface. This
is because libplacebo doesn't support multiplane textures today.

Note that the egl interop code also uses split planes so that's not
new behaviour.

For completeness, I'll note that the libplacebo logic is not using
VK_EXT_image_drm_format_modifier yet because it's not widely
available and so format negotiation is not done properly, and we're
lucky that things work out.