vo_gpu: vulkan: initialise vt_switcher and kms for display context
The value of doing this is a bit mixed. Hooking up the vt switcher
allows the video to pause on a vt switch, but there's another
problem.

The vulkan display code, at least on Intel, and probably AMD,
takes a drm lock that prevents kms code from restoring the crtc
mode as long as vulkan is active. That means we can't restore the
vt on a vt switch, so we can pause, but it'll keep showing the
paused video frame. Wonderful.

This locking also means that saving and restoring the mode after
the program exits is weird - we have to initialise kms to save
the mode, and then close our kms state so that Vulkan can start.

Then, at the end, after Vulkan closes, we have to reopen kms to
restore the mode. It actually works, but it's certainly odd.

For now, the drm device is being chosen independently of the vulkan
device. I've done some reading and the way you are supposed to do
this is via pci bus address matching (the UUID matching that we've
used elsewhere isn't supported). That will mean another libplacebo
change to expose the pci info.