vo_gpu: vulkan: drm and x11 integration for display context
While the basic Vulkan Display context can theoretically drive the
display without the involvement of any non-Vulkan code, the end result
is suboptimal:

* You can't save and restore the VT state
* You can't enable VAAPI interop
* You can't support the weird nvidia display stealing implementation

Solving the first two problems involves making direct drm calls like
we see in the egl-drm implementation, although it's a bit different
because we're initialising based on the device choice made by Vulkan
and we're not doing any mode setting except when restoring the mode
before we exit. Device matching is done by PCI bus ID.

And there is still a big problem:

The vulkan display code, at least on Intel, and probably AMD,
takes a drm lock that prevents kms code from restoring the crtc
mode as long as vulkan is active. That means we can't restore the
vt on a vt switch, so we can pause, but it'll keep showing the
paused video frame. Wonderful.

This locking also means that saving and restoring the mode after
the program exits is weird - we have to initialise kms to save
the mode, and then close our kms state so that Vulkan can start.

Then, at the end, after Vulkan closes, we have to reopen kms to
restore the mode. It actually works, but it's certainly odd.

To enable VAAPI interop, we need to open the render_fd, which is a
small change once the basic drm save/restore functionality is in
place. Luckily, we can keep the render_fd open though we have to
close the primary fd.

Finally, the nvidia implementation of VkDisplayKHR is naturally
completely different and of very limited use. I enabled it mostly just
to prove it can be done. The value here is dubious at best.

nvidia did not implement a standalone display mode. Instead, it works
by 'stealing' a display from X11 (doesn't work anywhere else) and then
giving dedicate Vulkan access to that display. The intention here was
to enable certain VR scenarios, and using this to show a fullscreen
video when you've still got X running on other displays is just silly.

It also doesn't work if there's just one display - you can't steal it.