x11: force VRR on
Rather than relying on the mesa driver to do this, add the atom
ourselves to make VRR kick in for nvidia render offload windows.

Of course, it doesn't work too well as the Intel GPU and/or Monitor
don't end up agreeing on the real frame rate that the nvidia GPU is
spitting out. Oh well.